+++ src/Infrastructure/Data/Repositories/CommentRepository.cs
@@ -19,7 +19,7 @@ namespace Mublog.Server.Infrastructure.Data.Repositories
             throw new System.NotImplementedException();
         }
 
-        public async Task<Comment> FindByIdAsync(int id)
+        public async Task<Comment> FindByIdAsync(long id)
         {
             throw new System.NotImplementedException();
         }
